+++ test/widgets/profile/gallery_test.dart
@@ -2,7 +2,7 @@ import 'package:MobileOne/localization/delegate.dart';
 import 'package:MobileOne/localization/supported.dart';
 import 'package:MobileOne/pages/profile.dart';
 import 'package:MobileOne/providers/user_picture_provider.dart';
-import 'package:MobileOne/services/authentication_service.dart';
+import 'package:MobileOne/providers/wishlistsList_provider.dart';
 import 'package:MobileOne/services/image_service.dart';
 import 'package:MobileOne/services/preferences_service.dart';
 import 'package:MobileOne/services/user_service.dart';
@@ -12,6 +12,7 @@ import 'package:flutter_localizations/flutter_localizations.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:get_it/get_it.dart';
 import 'package:google_sign_in/google_sign_in.dart';
+import 'package:image_picker/image_picker.dart';
 import 'package:mockito/mockito.dart';
 
 Widget buildTestableWidget(Widget widget) {
@@ -47,56 +48,60 @@ class MockGoogleSignIn extends Mock implements GoogleSignIn {}
 
 class FirebaseAuthMock extends Mock implements FirebaseAuth {}
 
-class PreferencesMock extends Mock implements PreferencesService {}
-
 class ImageServiceMock extends Mock implements ImageService {}
 
-class UserPictureMock extends Mock implements UserPictureProvider {}
+class PreferencesMock extends Mock implements PreferencesService {}
+
+class PickerMock extends Mock implements ImagePicker {}
+
+class WishlistListProviderMock extends Mock implements WishlistsListProvider {}
+
+class UserPictureProviderMock extends Mock implements UserPictureProvider {}
 
 void main() {
-  setSupportedLocales([Locale('fr', 'FR')]);
+  setSupportedLocales([Locale("fr", "FR")]);
 
-  testWidgets('Profile Page should display allow password modification',
+  testWidgets("Open gallery and see profile picture changed",
       (WidgetTester tester) async {
-    //GIVEN
-    Widget widget = buildTestableWidget(Profile());
-
-    final auth = FirebaseAuthMock();
+    // GIVEN
     final user = FirebaseUserMock();
-    final _googleSignIn = MockGoogleSignIn();
-    final _imageService = ImageServiceMock();
-    final _userService = UserServiceMock();
-    final _prefs = PreferencesMock();
-    final _picture = UserPictureMock();
-
-    GetIt.I.registerSingleton<UserService>(_userService);
-    GetIt.I.registerSingleton<GoogleSignIn>(_googleSignIn);
-    GetIt.instance.registerSingleton<FirebaseAuth>(auth);
     GetIt.instance.registerSingleton<FirebaseUser>(user);
-    GetIt.I.registerSingleton<AuthenticationService>(AuthenticationService());
+    final _userService = UserServiceMock();
+    GetIt.I.registerSingleton<UserService>(_userService);
+    final _googleSignIn = MockGoogleSignIn();
+    GetIt.I.registerSingleton<GoogleSignIn>(_googleSignIn);
+    final auth = FirebaseAuthMock();
+    GetIt.I.registerSingleton<FirebaseAuth>(auth);
+    final _prefService = PreferencesMock();
+    GetIt.I.registerSingleton<PreferencesService>(_prefService);
+    final _picker = PickerMock();
+    GetIt.I.registerSingleton<ImagePicker>(_picker);
+    final _imageService = ImageServiceMock();
     GetIt.I.registerSingleton<ImageService>(_imageService);
-    GetIt.I.registerSingleton<PreferencesService>(_prefs);
-    GetIt.I.registerSingleton<UserPictureProvider>(_picture);
+    final _wishlistProvider = WishlistListProviderMock();
+    GetIt.I.registerSingleton<WishlistsListProvider>(_wishlistProvider);
+    final _userPicture = UserPictureProviderMock();
+    GetIt.I.registerSingleton<UserPictureProvider>(_userPicture);
 
-    when(auth.currentUser()).thenAnswer((realInvocation) => Future.value(user));
-    final _displayName = "Dupond Jean";
-    when(user.displayName).thenReturn(_displayName);
+    when(_imageService.pickGallery()).thenAnswer((realInvocation) =>
+        Future.value(PickedFile("assets/images/facebook_f.png")));
 
     final _email = "Dupond.Jean@gmail.com";
     when(user.email).thenReturn(_email);
-
     when(user.photoUrl).thenReturn("");
 
+    when(_userService.user).thenReturn(user);
+    when(user.uid).thenReturn("45");
+    when(auth.currentUser()).thenAnswer((realInvocation) => Future.value(user));
     when(user.providerData).thenReturn(List.of([]));
 
-    when(_userService.user).thenReturn(user);
-
     //WHEN
-    await tester.pumpWidget(widget);
-    await tester.pumpAndSettle(Duration(seconds: 3));
-    await tester.tap(find.byKey(Key(KEY_PASSWORD)));
+    await tester.pumpWidget(buildTestableWidget(Profile()));
 
-    //THEN
-    expect(find.text('Changer mon mot de passe'), findsOneWidget);
+    await tester.pumpAndSettle(Duration(seconds: 1));
+    await tester.tap(find.byKey(Key(KEY_GALLERY)));
+
+    // THEN
+    verify(_prefService.setString("picture45", "assets/images/facebook_f.png"));
   });
 }